% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cblind_plot.R
\name{cblind_plot}
\alias{cblind_plot}
\title{Title}
\usage{
cblind_plot(
  im,
  r = 1,
  g = 2,
  b = 3,
  cvd = c("protanopia", "deuteranopia", "tritanopia"),
  crop_manual = FALSE,
  select_class = FALSE,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Title
}
\examples{
library(terra)
my_image <- rast(system.file("pic/rainbowr.png", package = "cblindplot"))
plotRGB(my_image)

cblind_plot(my_image)
#cblind_plot(my_image, crop_manual = TRUE)
#cblind_plot(my_image, select_class = TRUE)
}
