% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cblind_plot.R
\name{cblind_prep}
\alias{cblind_prep}
\title{Title}
\usage{
cblind_prep(im, r = 1, g = 2, b = 3, crop_manual = FALSE, select_class = FALSE)
}
\arguments{
\item{select_class}{}
}
\description{
Title
}
\examples{
library(terra)
my_image <- rast(system.file("pic/rainbowr.png", package = "cblindplot"))
plotRGB(my_image)

new_image <- cblind_prep(my_image)
new_image
}
