% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cblindplot-package.R
\docType{package}
\name{cblindplot-package}
\alias{cblindplot}
\alias{cblindplot-package}
\title{cblindplot: Recolors Images to be Colour Blind Friendly}
\description{
Enables colour blind people to enter an image, specify their colour vision deficiency, and get a colour blind friendly plot as output. It applies PCA analysis to the input image colors, and colors the values of the main principal component using a colour blind friendly palette.
}
\author{
\strong{Maintainer}: Elisa Thouverai \email{elisa.th95@gmail.com}

Authors:
\itemize{
  \item Duccio Rocchini \email{duccio.rocchini@unibo.it}
  \item Jakub Nowosad \email{nowosad.jakub@gmail.com} (\href{https://orcid.org/0000-0002-1057-3721}{ORCID})
}

}
\keyword{internal}
