% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cblind.plot.R
\name{cblind.plot}
\alias{cblind.plot}
\title{Color Blind Plot}
\usage{
cblind.plot(
  im,
  cvd = c("protanopia", "deuteranopia", "tritanopia"),
  r = 1,
  g = 2,
  b = 3,
  crop_manual = FALSE,
  select_class = FALSE
)
}
\arguments{
\item{im}{A file path to an image or a terra/raster object}

\item{cvd}{A type of color vision deficiency (CVD): "protanopia", "deuteranopia", or "tritanopia"}

\item{r}{Index of the Red channel}

\item{g}{Index of the Green channel}

\item{b}{Index of the Blue channel}

\item{crop_manual}{Do you want to manually crop the input image?}

\item{select_class}{Do you want to select only certain colors in the image for the further processing?}
}
\value{
A ggplot2 plot
}
\description{
It updates the color palette on an input image and returns a new visualization along with a new color legend.
}
\examples{
my_image <- system.file("pic/imager.png", package = "cblindplot")

#suppressWarnings(my_image_terra <- terra::rast(my_image))
#terra::plotRGB(my_image_terra)

cblind.plot(my_image)
#cblind.plot(my_image, crop_manual = TRUE)
#cblind.plot(my_image, select_class = TRUE)
}
